refactor: extract bot AI into server/bot.js module
